der: consolidate `bytes`/`string` modules
Elsewhere we have single modules containing both reference types and
their equivalent allocating types.

This follows the same pattern.

Also renames (non-breaking, internal types);
- `StrRef` => `StringRef`
- `StrOwned` => `StringOwned`
