+++ der/src/asn1/printable_string.rs
@@ -1,6 +1,6 @@
 //! ASN.1 `PrintableString` support.
 
-use crate::{FixedTag, Result, StrRef, Tag, asn1::AnyRef};
+use crate::{FixedTag, Result, StringRef, Tag, asn1::AnyRef};
 use core::{fmt, ops::Deref};
 
 macro_rules! impl_printable_string {
@@ -54,7 +54,7 @@ macro_rules! impl_printable_string {
 #[derive(Copy, Clone, Eq, PartialEq, PartialOrd, Ord)]
 pub struct PrintableStringRef<'a> {
     /// Inner value
-    inner: StrRef<'a>,
+    inner: StringRef<'a>,
 }
 
 impl<'a> PrintableStringRef<'a> {
@@ -87,7 +87,7 @@ impl<'a> PrintableStringRef<'a> {
             }
         }
 
-        StrRef::from_bytes(input)
+        StringRef::from_bytes(input)
             .map(|inner| Self { inner })
             .map_err(|_| Self::TAG.value_error().into())
     }
@@ -96,7 +96,7 @@ impl<'a> PrintableStringRef<'a> {
 impl_printable_string!(PrintableStringRef<'a>, 'a);
 
 impl<'a> Deref for PrintableStringRef<'a> {
-    type Target = StrRef<'a>;
+    type Target = StringRef<'a>;
 
     fn deref(&self) -> &Self::Target {
         &self.inner
@@ -122,7 +122,7 @@ mod allocation {
     use super::PrintableStringRef;
 
     use crate::{
-        BytesRef, Error, FixedTag, Result, StrOwned, Tag,
+        BytesRef, Error, FixedTag, Result, StringOwned, Tag,
         asn1::AnyRef,
         referenced::{OwnedToRef, RefToOwned},
     };
@@ -159,7 +159,7 @@ mod allocation {
     #[derive(Clone, Eq, PartialEq, PartialOrd, Ord)]
     pub struct PrintableString {
         /// Inner value
-        inner: StrOwned,
+        inner: StringOwned,
     }
 
     impl PrintableString {
@@ -171,7 +171,7 @@ mod allocation {
             let input = input.as_ref();
             PrintableStringRef::new(input)?;
 
-            StrOwned::from_bytes(input)
+            StringOwned::from_bytes(input)
                 .map(|inner| Self { inner })
                 .map_err(|_| Self::TAG.value_error().into())
         }
@@ -180,7 +180,7 @@ mod allocation {
     impl_printable_string!(PrintableString);
 
     impl Deref for PrintableString {
-        type Target = StrOwned;
+        type Target = StringOwned;
 
         fn deref(&self) -> &Self::Target {
             &self.inner
@@ -190,7 +190,7 @@ mod allocation {
     impl<'a> From<PrintableStringRef<'a>> for PrintableString {
         fn from(value: PrintableStringRef<'a>) -> PrintableString {
             let inner =
-                StrOwned::from_bytes(value.inner.as_bytes()).expect("Invalid PrintableString");
+                StringOwned::from_bytes(value.inner.as_bytes()).expect("Invalid PrintableString");
             Self { inner }
         }
     }
@@ -234,7 +234,7 @@ mod allocation {
         fn try_from(input: String) -> Result<Self> {
             PrintableStringRef::new(&input)?;
 
-            StrOwned::new(input)
+            StringOwned::new(input)
                 .map(|inner| Self { inner })
                 .map_err(|_| Self::TAG.value_error().into())
         }
